requirements.txt
torch
matplotlib
Pillow
torchvision
